+++ CHANGELOG.md
@@ -1,12 +1,21 @@
+v0.1.5 / Mar 30 2015
+=========================
+Add Angular options by constant `jaggyConfig`
+
+ * [`unknown`][2] :bulb: empty image instead of Error by `jaggyConfig.useEmptyImage`
+ * [`unknown`][2] :bulb: caching a converted svg by `jaggyConfig.useCache`
+
+[2]: https://github.com/59naga/jaggy/commits/master
+
 v0.1.4 / Mar 29 2015
 =========================
- * [`unknown`][1] :bug: Fix `<path fill="rgba(,,,NaN)"`
- * [`unknown`][1] :lipstick: Support `<img ng-src="" jaggy>`
+ * [`21fb96a`][1] :bug: Fix `<path fill="rgba(,,,NaN)"`
+ * [`21fb96a`][1] :lipstick: Support `<img ng-src="" jaggy>`
 
-[1]: https://github.com/59naga/jaggy/commit
+[1]: https://github.com/59naga/jaggy/commit/21fb96a22352c84f4802c50f6a35f7500cee9254
 
 v0.1.3 / Mar 03 2015
 =========================
- * [`unknown`][0] Release v0.1.3
+ * [`1084961c`][0] Release v0.1.3
 
 [0]: https://github.com/59naga/jaggy/commits/master